Add caching device for memory.
Add a cache simulator to Pep/9.
For the moment, it does not have a visualization component.

It maintains efficiency by storing references to memory rather than duplicating memory contents.
This feature is allowed because we use write-through caching rather than write-back.
Write-through is mandatory for the Pep/9 processor because of memory mapped IO.
Otherwise, repeated writes to the same memory address may be elided, which would "lose" MMIO operations.

Allowing write-back would require that MMIO addresses be treated differently in the cache than non-MMIO addresses.
While this may be required in silicon, no need to bring this level of complexity to the Pep/9 project.