Add ability to exclude OS instructions from statistics computations.
The statistics widget now has a checkbox to include/exclude the contributions of the OS to execution of the program.
If data is present in the statistics widget, the mode may be changed dynamically by changing the check state of the box.

This required changing how statistics are computed / stored in ISA, microcode models.