Perform eviction collation, pad cache address to 4 characters.
Addresses in the cache are all padded with zeros to a length of 4.

When an index has been evicted multiple times during a single simulation step (like during a trap instruction), only the most recent eviction for each simulation step is kept.
Otherwise, dozens to hundreds of eviction entries may flood the screen.
This behavior may change in the future, so it may currently be enabled or disabled at compile time using a if constexpr.